16 Remove slf4j-api from the Server’s runtime.
Inside a container, the SLF4J API needs to be loaded from the same
class loader as its implementation. So either we add an implementation
as a runtime dependency, either we remove the API from the runtime
dependencies.

I choosed to do the latter because it allows the user to choose which
implementation to use at runtime. The downside is that it requires
them to add jars in their container.